refactor everything out of lib/__init__.py
