+++ lib/extensions/embroider.py
@@ -1,12 +1,13 @@
 import sys
 import traceback
 import os
-
 import inkex
-from .. import _, PIXELS_PER_MM, write_embroidery_file
+
 from .base import InkstitchExtension
+from ..i18n import _
+from ..output import write_embroidery_file
 from ..stitch_plan import patches_to_stitch_plan
-from ..svg import render_stitch_plan
+from ..svg import render_stitch_plan, PIXELS_PER_MM
 
 
 class Embroider(InkstitchExtension):
