+++ lib/stitch_plan/__init__.py
@@ -1,1 +1,2 @@
 from stitch_plan import patches_to_stitch_plan, StitchPlan, ColorBlock
+from .stitch import Stitch
